backend/complaint_system/requirements.txt
Django>=4.2.0
django-crispy-forms>=2.0
Pillow>=9.0.0
python-dotenv>=1.0.0
djangorestframework>=3.14.0
django-filter>=23.2
django-cors-headers>=4.2.0
django-allauth>=0.54.0
django-export-download>=1.1.0
django-simple-history>=3.3.0
django-widget-tweaks>=1.5.0
psycopg2-binary>=2.9.7
whitenoise>=6.5.0
gunicorn>=21.2.0
django-bootstrap4>=23.2
xlwt>=1.3.0
reportlab>=4.0.4
